requirements.txt
python-binance>=1.0.20
pandas>=2.1.4
numpy>=1.26.2
ta>=0.11.0
python-dotenv>=1.0.0
ccxt>=4.1.13
scikit-learn>=1.3.2
matplotlib>=3.8.2
pytest>=7.4.3 
pydantic>=2.10.2